change routes structure
