+++ app/routes/drawer/routes/reactWall/reactWallScene.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { View, Text, StyleSheet } from 'react-native';
 import { COLOR, ThemeProvider, Toolbar } from 'react-native-material-ui';
-import Container from '../drawer/container';
-import Header from '../../components/header.js';
-
+import Container from '../../container';
+import Icon from 'react-native-vector-icons/Ionicons';
+import Header from '../../../../components/header.js';
 const uiTheme = {
   palette: {
     primaryColor: COLOR.grey500,
@@ -11,16 +11,17 @@ const uiTheme = {
   },
 };
 
-class SettingsScene extends Component {
+class reactWallSceen extends Component {
   render () {
+    const { errorToken } = this.props;
     return (
           <ThemeProvider uiTheme={uiTheme}>
             <Container>
               <Header navigatorLeft={ () => this.props.navigation.navigate('DrawerOpen')}
-                      title="React Settings"/>
+                      title="React Wall"/>
               <View style={styles.container}>
                 <Text style={styles.welcome}>
-                  Welcome to Settings
+                  Welcome to ReactWall!
                 </Text>
               </View>
             </Container>
@@ -43,4 +44,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default SettingsScene;
+export default reactWallSceen;
